requirements.txt
anyio==4.4.0
black==24.4.2
blinker==1.8.2
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
docutils==0.21.2
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
lxml_html_clean==0.1.1
marimo==0.6.25
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psutil==6.0.0
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
readability-lxml==0.8.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
usearch==2.12.0
uvicorn==0.30.1
websockets==12.0
Werkzeug==3.0.3
